1. bug fix in comm.hpp: isend must use heap variable, to be tested. 2. bug fix in mf.hpp: register_bupdate using sync 3. bug to fix:  update function with reference

+++ alg/matrix_factorization/update.cpp
@@ -26,7 +26,7 @@ extern "C" {
 
 vector<double> local_update_fac(vector<double> a, vector<double> b) {
   vector<double> r;
-  for(int i = 0; i < (int)a.size(); ++i) {
+  for(size_t i = 0; i < a.size(); ++i) {
     r.push_back(a[i] + b[i]);
   }
   return r;
